+++ changelog.md
@@ -1,4 +1,5 @@
 #### v1.0.1
 - looks for samplesheet with case insensitive matching of glob(data/run/*csv) items containing "samplesheet" or "data"
 - removes carriage return: ^M from samplesheet which occurs when user copy and pastes sample IDs from excel into SPY's samplesheet creator.
-- NEED TO add refresh to list runs...
+- Button added to refresh run listing
+- Downloaded fasta for FLU now changes the flu segment numbers to the gene name